refactor: move from tasks to packages
Переместил все файлы из `tasks` в `packages`
