+++ jest.unit.config.js
@@ -1,11 +1,7 @@
 const path = require("path");
 module.exports = {
   displayName: "unit",
-  roots: [
-    path.join(__dirname, "src"),
-    path.join(__dirname, "tasks"),
-    path.join(__dirname, "postcss-plugin-vk-sans"),
-  ],
+  roots: [path.join(__dirname, "src"), path.join(__dirname, "packages")],
   setupFilesAfterEnv: [path.join(__dirname, "src/testing/setup.ts")],
   collectCoverage: true,
   collectCoverageFrom: ["src/*/**/*.{ts,tsx}"],
