+++ packages/eslint-plugin-vkui/package.json
