+++ packages/postcss-plugin-vk-sans/__tests__/custom_properties.css
