+++ packages/stylelint-atomic/index.js
