+++ packages/stylelint-bad-multiplication/index.js
